+++ pycomex/examples/quickstart.py
@@ -3,6 +3,7 @@
 This doc string will be saved as the "description" meta data of the experiment records
 """
 from pycomex.experiment import Experiment
+from pycomex.util import Skippable
 
 # Experiment parameters can simply be defined as uppercase global variables.
 # These are automatically detected and can possibly be overwritten in command
@@ -14,8 +15,7 @@ WORLD = "world!"
 # - Path to an existing folder in which to store the results
 # - A namespace name unique for each experiment
 # - access to the local globals() dict
-with Experiment("/tmp", "example/quickstart", globals()) as e:
-    e.prepare()  # important!
+with Skippable(), (e := Experiment("/tmp", "example/quickstart", globals())):
 
     # Internally saved into automatically created nested dict
     # {'strings': {'hello_world': '...'}}
@@ -32,12 +32,12 @@ with Experiment("/tmp", "example/quickstart", globals()) as e:
     # e.commit_png(file_name, image)
     # ...
 
-    # All the code inside this context will be copied to the "analysis.py"
-    # file which will be created as an experiment artifact.
-    with e.analysis:
-        # And we can access all the internal fields of the experiment object
-        # and the experiment parameters here!
-        print(HELLO, WORLD)
-        print(e['strings/hello_world'])
-        # logging will print to stdout but not modify the log file
-        e.info('analysis done')
+# All the code inside this context will be copied to the "analysis.py"
+# file which will be created as an experiment artifact.
+with Skippable(), e.analysis:
+    # And we can access all the internal fields of the experiment object
+    # and the experiment parameters here!
+    print(HELLO, WORLD)
+    print(e['strings/hello_world'])
+    # logging will print to stdout but not modify the log file
+    e.info('analysis done')